Application is up and running.
* The application is now working flawlessly.
* The UI got a slight upgrade, now it uses tabs in the main menu.
* Removed redundant files.

+++ Source/Route.cpp
@@ -37,7 +37,7 @@ const int Route::GetTotalItemCount() const
 {
 	int count = 0;
 	for (const Order order : mOrders)
-		count += order.mPackages.size();
+		count += static_cast<int>(order.mPackages.size());
 
 	return count;
 }
